+++ src/transformers/models/qwen3_5_moe/modeling_qwen3_5_moe.py
@@ -49,6 +49,7 @@ from ...processing_utils import Unpack
 from ...utils import TransformersKwargs, auto_docstring, can_return_tuple, logging, torch_compilable_check
 from ...utils.deprecation import deprecate_kwarg
 from ...utils.generic import (
+    accelerate_hook_compatible_wrapper,
     accepts_precomputed_kwargs,
     is_flash_attention_requested,
     maybe_autocast,
@@ -364,37 +365,6 @@ def torch_recurrent_gated_delta_rule(
         last_recurrent_state = None
     core_attn_out = core_attn_out.transpose(1, 2).contiguous().to(initial_dtype)
     return core_attn_out, last_recurrent_state
-
-
-def accelerate_hook_compatible_wrapper(original_func: Callable) -> Callable:
-    """
-    Wrapper around a function to trigger the accelerate hooks even when the weights are used directly rather than
-    through their `forward`, as is the case inside `causal_conv1d_fn` and `causal_conv1d_update`.
-    """
-
-    def wrapped(*args, **kwargs):
-        hook = None
-        hooked_module = kwargs.pop("hooked_module", None)
-        if hooked_module is not None:
-            if (hook := getattr(hooked_module, "_hf_hook", None)) is not None:
-                # args, kwargs = hook.pre_forward(hooked_module, *args, **kwargs)
-                hook.pre_forward(hooked_module)
-
-        # Since the weights of the module were passed to the caller before being moved, we need to re-update them
-        if hook is not None:
-            if "weight" in kwargs:
-                kwargs["weight"] = hooked_module.weight.squeeze(1)
-            if "bias" in kwargs:
-                kwargs["bias"] = hooked_module.bias
-
-        output = original_func(*args, **kwargs)
-
-        if hook is not None:
-            return hook.post_forward(hooked_module, output)
-        else:
-            return output
-
-    return wrapped
 
 
 @use_kernel_forward_from_hub("Qwen3_5GatedDeltaNet")
